Appointment_Service/requirements.txt
Flask==2.0.3
psycopg2
python-dotenv==1.0.0
